+++ app/src/androidTest/java/net/sourceforge/opencamera/test/MainActivityTest.java
@@ -16331,7 +16331,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_y = 47.44656f;
         float panorama_pics_per_screen = 2.0f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (47.44656/49.56283);
+        panorama_pics_per_screen *= (float)(47.44656/49.56283);
         String output_name = "testPanorama1_output.jpg";
 
         subTestPanorama(inputs, output_name, null, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 2.0f);
@@ -16368,7 +16368,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, null, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 2.0f);
     }
@@ -16399,7 +16399,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, null, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 1.0f);
     }
@@ -16431,7 +16431,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, null, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 1.0f);
     }
@@ -16460,7 +16460,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 1.0f);
     }
@@ -16489,7 +16489,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16518,7 +16518,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16548,7 +16548,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16573,7 +16573,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/52.26029);
+        panorama_pics_per_screen *= (float)(50.282097/52.26029);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16601,7 +16601,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/50.44399);
+        panorama_pics_per_screen *= (float)(50.282097/50.44399);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
 
@@ -16638,7 +16638,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/50.44399);
+        panorama_pics_per_screen *= (float)(50.282097/50.44399);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16666,7 +16666,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/50.44399);
+        panorama_pics_per_screen *= (float)(50.282097/50.44399);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
@@ -16697,7 +16697,7 @@ public class MainActivityTest extends ActivityInstrumentationTestCase2<MainActiv
         float camera_angle_x = 66.708595f;
         float camera_angle_y = 50.282097f;
         // these images were taken with incorrect camera view angles, so we compensate in the test:
-        panorama_pics_per_screen *= (50.282097/50.44399);
+        panorama_pics_per_screen *= (float)(50.282097/50.44399);
 
         subTestPanorama(inputs, output_name, gyro_name, panorama_pics_per_screen, camera_angle_x, camera_angle_y, 0.5f);
     }
